split out classes into separate modules, for better compatibility with memory constrained systems
The 1115 is the most prevalent ADC, and ADS1115 is the base class
anyway. So this is left in the original module.

The original module contains compatibility wrappers so that import
statements in user code do not have to be modified.

constants cannot be imported so some of them need to be duplicated
